+++ client/src/Components/UserHeader.tsx
@@ -5,9 +5,9 @@ import { HTMLAttributes, useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import styled from 'styled-components';
 
+import useAxios from '../apis/customAxios';
 import { Iuser } from '../reducers/userInfoReducer';
 import { RootState } from '../store/Store';
-import useAxios from '../util/customAxios';
 
 interface UserHeaderImgProps extends HTMLAttributes<HTMLDivElement> {
 	image?: string;
