+++ client/src/pages/HotReview.tsx
@@ -3,7 +3,7 @@ import { HTMLAttributes, useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 
-import useAxios from '../util/customAxios';
+import useAxios from '../apis/customAxios';
 
 interface SlideItemProps extends HTMLAttributes<HTMLDivElement> {
 	image?: string;
