+++ client/src/pages/member/Mypage.tsx
@@ -7,12 +7,12 @@ import { Link, useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 import Swal from 'sweetalert2';
 
+import { Api } from '../../apis/customAPI';
+import useAxios from '../../apis/customAxios';
 import IntroBox from '../../Components/member/IntroBox';
 import MyReview from '../../Components/member/MyReview';
 import { Iuser, UPDATE } from '../../reducers/userInfoReducer';
 import { RootState } from '../../store/Store';
-import { Api } from '../../util/customAPI';
-import useAxios from '../../util/customAxios';
 
 const Main = styled.div`
 	display: flex;
